pyproject.toml
[tool.poetry]
name = "pyconau-2020-website"
version = "0.1.0"
description = ""
authors = ["Leigh Brenecki <leigh@brenecki.id.au>"]

[tool.poetry.dependencies]
python = "^3.7"
statik = "^0.23.0"
sqlalchemy = {version = "*", extras = []}
"ruamel.yaml" = "^0.16.10"
parse = "^1.16.0"

[tool.poetry.dev-dependencies]
supervisor = "^4.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
